+++ SCEvent.m
@@ -157,7 +157,7 @@
 // -------------------------------------------------------------------------------
 - (NSString *)description
 {
-    return [NSString stringWithFormat:@"<%@ { eventId = %d, eventPath = %@, eventFlag = %d } >", [self className], _eventId, _eventPath, _eventFlag];
+    return [NSString stringWithFormat:@"<%@ { eventId = %d, eventPath = %@, eventFlag = %d } >", [self className], (int)_eventId, _eventPath, (int)_eventFlag];
 }
 
 // -------------------------------------------------------------------------------
